+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.example.retrofit_"
+        applicationId "com.example.retrofit"
         minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
