+++ app/src/androidTest/java/com/example/retrofit/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.retrofit_
+package com.example.retrofit
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
