+++ app/src/main/java/com/example/retrofit/Data.kt
@@ -1,4 +1,4 @@
-package com.example.retrofit_
+package com.example.retrofit
 
 import retrofit2.http.GET
 import retrofit2.http.Query
