+++ app/src/main/java/com/example/retrofit/DataBean.kt
@@ -1,4 +1,4 @@
-package com.example.retrofit_
+package com.example.retrofit
 
 data class DataBean(
